[General] Fixed many obvious warnings (went from 300+ down to 124).

+++ sources/tools/SiliconStudio.Paradox.Importer.Assimp/Paradox.Importer.Assimp.cpp
@@ -1067,7 +1067,7 @@ private:
 
 			auto itemPart = std::string();
 
-			int itemNameSplitPosition = itemName.find('#');
+			size_t itemNameSplitPosition = itemName.find('#');
 			if (itemNameSplitPosition != std::string::npos)
 			{
 				itemPart = itemName.substr(itemNameSplitPosition + 1);
@@ -1171,9 +1171,9 @@ private:
 		GetNodeNames(scene->mRootNode, baseNames, orderedNodes);
 
 		// Need to create the array of the nodes
-		int nodeCount = orderedNodes.size();
+		size_t nodeCount = orderedNodes.size();
 		aiNode** nodeArray = new aiNode*[nodeCount];
-		for (int i = 0; i < nodeCount; ++i)
+		for (size_t i = 0; i < nodeCount; ++i)
 			nodeArray[i] = orderedNodes[i];
 
 		GenerateUniqueNames(nodeNames, baseNames, nodeArray);
