+++ sources/tools/SiliconStudio.Paradox.Importer.Common/ImporterUtils.h
@@ -5,7 +5,7 @@
 
 void ReplaceCharacter(std::string& name, char c, char replacement)
 {
-	int nextCharacterPos = name.find(c);
+	size_t nextCharacterPos = name.find(c);
 	while (nextCharacterPos != std::string::npos)
 	{
 		name.replace(nextCharacterPos, 1, 1, replacement);
@@ -15,7 +15,7 @@ void ReplaceCharacter(std::string& name, char c, char replacement)
 
 void RemoveCharacter(std::string& name, char c)
 {
-	int nextCharacterPos = name.find(c);
+	size_t nextCharacterPos = name.find(c);
 	while (nextCharacterPos != std::string::npos)
 	{
 		name.erase(nextCharacterPos, 1);
